+++ utility/src/main/java/com/example/entity/dto/ImportProductDetail.java
@@ -13,13 +13,13 @@ import jakarta.persistence.Table;
 public class ImportProductDetail {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
+  private Integer id;
 
-  public Long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
 
@@ -47,13 +47,13 @@ public class ImportProductDetail {
     this.importProduct = importProduct;
   }
 
-  private Long quantity;
+  private Integer quantity;
 
-  public Long getQuantity() {
+  public Integer getQuantity() {
     return this.quantity;
   }
 
-  public void setQuantity(Long quantity) {
+  public void setQuantity(Integer quantity) {
     this.quantity = quantity;
   }
 }
